unix-echo -> unix-echo-server; int -> size_t in array

+++ array_heap.c
@@ -5,7 +5,7 @@
 
 int array_init(array* arr, int size) {
   arr->data = realloc(NULL, sizeof(void*) * size);
-  if (0 == (int)arr->data) {
+  if (0 == (size_t)arr->data) {
     return -1;
   }
   arr->length = size;
@@ -14,7 +14,7 @@ int array_init(array* arr, int size) {
 }
 
 int array_push(array* arr, void* data) {
-  ((int*)arr->data)[arr->index] = (int)data;
+  ((size_t*)arr->data)[arr->index] = (size_t)data;
   arr->index += 1;
   if (arr->index >= arr->length) {
     if (-1 == array_grow(arr, arr->length * 2))
@@ -30,7 +30,7 @@ int array_grow(array* arr, int size) {
     return -1;
   }
   arr->data = realloc(arr->data, sizeof(void*) * size);
-  if (-1 == (int)arr->data) {
+  if (-1 == (size_t)arr->data) {
     return -1;
   }
   arr->length = size;
@@ -40,7 +40,7 @@ int array_grow(array* arr, int size) {
 void array_free(array* arr, void (*free_element)(void*)) {
   int i;
   for (i = 0; i < arr->index; i += 1) {
-    free_element((void*)((int*)arr->data)[i]);
+    free_element((void*)((size_t*)arr->data)[i]);
   }
   free(arr->data);
   arr->index = -1;
